feat: profiles controller

+++ app/Http/Controllers/DepartamentoController.php
@@ -3,7 +3,9 @@
 namespace App\Http\Controllers;
 
 use App\Models\Departamento;
+use App\Models\User;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class DepartamentoController extends Controller
 {
@@ -14,6 +16,14 @@ class DepartamentoController extends Controller
      */
     public function index()
     {
+        $user = User::findOrFail(Auth::id());
+        if(!$user->can('listar departamento') && !$user->hasRole('super-admin')){
+            return view('auth.unauthorized', [
+                'root' => 'Departamentos',
+                'page' => '',
+            ]);
+        }
+        
         $departamentos = Departamento::all();
         // $departamentos = Departamento::withTrashed()->get();
         return view('departamentos.index', [
@@ -30,6 +40,14 @@ class DepartamentoController extends Controller
      */
     public function create()
     {
+        $user = User::findOrFail(Auth::id());
+        if(!$user->can('crear departamento') && !$user->hasRole('super-admin')){
+            return view('auth.unauthorized', [
+                'root' => 'Departamentos',
+                'page' => '',
+            ]);
+        }
+        
         return view('departamentos.create', [
             'root' => 'Departamentos',
             'page' => 'Crear',
@@ -44,6 +62,13 @@ class DepartamentoController extends Controller
      */
     public function store(Request $request)
     {
+        $user = User::findOrFail(Auth::id());
+        if(!$user->can('crear departamento') && !$user->hasRole('super-admin')){
+            return view('auth.unauthorized', [
+                'root' => 'Departamentos',
+                'page' => '',
+            ]);
+        }
         $request->validate([
             'nombre_departamento' => 'bail|required',
         ]);
@@ -66,6 +91,13 @@ class DepartamentoController extends Controller
      */
     public function show($id)
     {
+        $user = User::findOrFail(Auth::id());
+        if(!$user->can('ver departamento') && !$user->hasRole('super-admin')){
+            return view('auth.unauthorized', [
+                'root' => 'Departamentos',
+                'page' => '',
+            ]);
+        }
         $departamento = Departamento::findOrFail($id);
         return view('departamentos.show', [
             'root' => 'Departamentos',
@@ -82,6 +114,13 @@ class DepartamentoController extends Controller
      */
     public function edit($id)
     {
+        $user = User::findOrFail(Auth::id());
+        if(!$user->can('editar departamento') && !$user->hasRole('super-admin')){
+            return view('auth.unauthorized', [
+                'root' => 'Departamentos',
+                'page' => '',
+            ]);
+        }
         $departamento = Departamento::findOrFail($id);
         return view('departamentos.edit', [
             'root' => 'Departamentos',
@@ -99,6 +138,13 @@ class DepartamentoController extends Controller
      */
     public function update(Request $request, $id)
     {
+        $user = User::findOrFail(Auth::id());
+        if(!$user->can('editar departamento') && !$user->hasRole('super-admin')){
+            return view('auth.unauthorized', [
+                'root' => 'Departamentos',
+                'page' => '',
+            ]);
+        }
         $departamento = Departamento::findOrFail($id);
         
         $request->validate([
@@ -122,6 +168,13 @@ class DepartamentoController extends Controller
      */
     public function destroy($id)
     {
+        $user = User::findOrFail(Auth::id());
+        if(!$user->can('borrar departamento') && !$user->hasRole('super-admin')){
+            return view('auth.unauthorized', [
+                'root' => 'Departamentos',
+                'page' => '',
+            ]);
+        }
         $departamento = Departamento::findOrFail($id);
         $departamento->delete();
         return redirect('/departamentos');
